Move mock LLM client out of agent runner
